enemigo esquiva y crosshair

+++ Assets/Scripts/SOBase/WeaponScriptable.cs
@@ -25,7 +25,7 @@ public class WeaponScriptable : ScriptableObject
     #region SpecialShoot
     [Header("Special Shoot")]
     public SpecialAttack SpecialType;
-    public float SpecialDamage;
+    public int SpecialDamage;
     public float SpecialStartup;
     public float SpecialTime;
     public float SpecialCooldown;
